# apispecs/src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Web Project API
  description: Api connect to Database
  version: 0.0.1
  license:
    name: TDTU Web Project
servers:
  - url: http://XXXXX/v1
    description: Develop server
  - url: http://XXXXX.com
    description: Product server
paths:
...
  /v1/createNews:
    post:
      tags:
        - News
      summary: Api for create News
      operationId: createNews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCreateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsCreateResponse'
  /v1/updateNews:
    post:
      tags:
        - News
      summary: Api for update News
      operationId: updateNews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsUpdateResponse'
  /v1/deleteNews:
    delete:
      tags:
        - News
      summary: Api for delete news
      operationId: deleteNews
      parameters:
        - name: newsId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status delete news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsDeleteResponse'
  /v1/searchNews:
    post:
      tags:
        - News
      summary: Api for search news
      operationId: searchNews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsSearchRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Search news in data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsSearchResponse'
  /v1/getNews:
    get:
      tags:
        - News
      summary: Api for get News by NewsId
      operationId: getNews
      parameters:
        - name: newsId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsGetResponse'
...
components:
...
  schemas:
    WebApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/WebApiErrorDetails'
    WebApiErrorDetails:
      type: object
      properties:
        field:
          type: string
        errorCd:
          type: string
        message:
          type: string
        errorIdList:
          type: array
          items:
            type: integer
            example: 1
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string
...
    NewsCreateRequest:
      type: object
      required:
        - title
        - content
        - createUserId
      properties:
        title:
          description: Title of news
          type: string
          example: 'xxxxx'
          maxLength: 4000
        content:
          description: Content of news
          type: string
          example: 'xxxxx'
          maxLength: 4000
        createUserId:
          description: ID of user create news
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        cover:
          description: Link of cover
          type: string
          example: 'xxxxx'
          maxLength: 200
    NewsCreateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create news
          type: string
          example: 'OK'
    NewsUpdateRequest:
      type: object
      required:
        - newsId
        - title
        - content
        - lastupUserId
      properties:
        newsId:
          description: ID of news
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        title:
          description: Title of news
          type: string
          example: 'xxxxx'
          maxLength: 4000
        content:
          description: Content of news
          type: string
          example: 'xxxxx'
          maxLength: 4000
        lastupUserId:
          description: ID of user update news
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        cover:
          description: Link of cover
          type: string
          example: 'xxxxx'
          maxLength: 200
    NewsUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update news
          type: string
          example: 'OK'
    NewsDeleteResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the delete news
          type: string
          example: 'OK'
    NewsSearchRequest:
      x-custom-validate:
        date-range:
          - from: 'createDatetimeFrom'
            to: 'createDatetimeTo'
      type: object
      properties:
        title:
          description: Title of news
          type: string
          example: 'xxxxx'
          maxLength: 4000
        userName:
          description: Name of user
          type: string
          example: 'user'
          maxLength: 200
        createDatetimeFrom:
          description: Date time create news
          type: string
          example: '123456'
          pattern: '^([12]\d{3})\/(0[1-9]|1[0-2])\/(0[1-9]|1\d|2[0-8]|29(?=\/\d\d\/(?!1[01345789]00|2[1235679]00)\d\d(?:[02468][048]|[13579][26]))|30(?!\/02)|31(?=\/0[13578]|\/1[02]))$'
        createDatetimeTo:
          description: Date time create news
          type: string
          example: '123456'
          pattern: '^([12]\d{3})\/(0[1-9]|1[0-2])\/(0[1-9]|1\d|2[0-8]|29(?=\/\d\d\/(?!1[01345789]00|2[1235679]00)\d\d(?:[02468][048]|[13579][26]))|30(?!\/02)|31(?=\/0[13578]|\/1[02]))$'
        offset:
          description: The offset rows before beginning to return the rows
          type: integer
          format: int32
          minimum: 0
          example: 1
        limit:
          description: The number of rows
          type: integer
          format: int32
          minimum: 1
          example: 20
    NewsSearchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            resultsTotalCount:
              description: Number of new with condition
              type: integer
              format: int64
              example: 1000
            news:
              description: News list
              type: array
              items:
                $ref: '#/components/schemas/NewsResponse'
    NewsResponse:
      type: object
      properties:
        newsId:
          description: ID of news
          type: string
          example: '123456'
        title:
          description: Title of news
          type: string
          example: '123456'
        content:
          description: Content of news
          type: string
          example: 'xxxxx'
        cover:
          description: Cover of news
          type: string
          example: '123456'
        createUserId:
          description: ID of user create news
          type: string
          example: '123456'
        createUserName:
          description: Name of user create news
          type: string
          example: 'xxxxx'
        createDatetime:
          description: Date-time create news
          type: string
          example: '2022-12-20 00:00:00'
        lastupUserId:
          description: ID of user lastup news
          type: string
          example: '123456'
        lastupUserName:
          description: Name of user lastup news
          type: string
          example: 'xxxxx'
        lastupDatetime:
          description: Date-time lastup news
          type: string
          example: '2022-12-20 00:00:00'
    NewsGetResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/NewsResponse'
...
  responses:
    webApiValidationError:
      description: webApi input value validation failure
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/WebApiErrorResponse'
            properties:
              status:
                example: 400
...
    401:
      description: Restrict access
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 401
              error:
                example: Unauthorized
    403:
      description: Access request denied
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 403
              error:
                example: Forbidden
    404:
      description: Page not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 404
              error:
                example: Not Found
    500:
      description: The server is having problems
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 500
              error:
                example: Internal Server Error
...